+++ test/web/twitter_api/views/activity_view_test.exs
@@ -91,16 +91,16 @@ defmodule Pleroma.Web.TwitterAPI.ActivityViewTest do
   test "a create activity with a summary containing emoji" do
     {:ok, activity} =
       CommonAPI.post(insert(:user), %{
-        "spoiler_text" => ":woollysocks: meow",
+        "spoiler_text" => ":firefox: meow",
         "status" => "."
       })
 
     result = ActivityView.render("activity.json", activity: activity)
 
-    expected = ":woollysocks: meow"
+    expected = ":firefox: meow"
 
     expected_html =
-      "<img height=\"32px\" width=\"32px\" alt=\"woollysocks\" title=\"woollysocks\" src=\"http://localhost:4001/finmoji/128px/woollysocks-128.png\" /> meow"
+      "<img height=\"32px\" width=\"32px\" alt=\"firefox\" title=\"firefox\" src=\"http://localhost:4001/emoji/Firefox.gif\" /> meow"
 
     assert result["summary"] == expected
     assert result["summary_html"] == expected_html